Revert "[DRAFT] MLSys RAG project implementation in MLC-LLM"

+++ python/mlc_llm/model/gpt2/gpt2_model.py
@@ -28,7 +28,7 @@ class GPT2Config(ConfigBase):  # pylint: disable=too-many-instance-attributes
     n_embd: int
     n_layer: int
     n_head: int
-    layer_norm_epsilon: float
+    layer_norm_epsilon: int
     n_inner: int = -1
     context_window_size: int = 0
     prefill_chunk_size: int = 0
